If block has no assignees or labels, dont make query to remove them

+++ backend/blueprints/blocks.py
@@ -90,17 +90,18 @@ def remove_block_route(block_id: int, token_id: int):
     task_ids = [task_row['id'] for task_row in task_rows]
 
     # Creating delete query for block's tasks
-    where_clause = ' or '.join([f'task_id = {task_id}' for task_id in task_ids])
-    assignee_query = f"DELETE FROM assignees WHERE {where_clause}"
-    label_query = f"DELETE FROM task_labels WHERE {where_clause}"
+    if len(task_ids):
+        where_clause = ' or '.join([f'task_id = {task_id}' for task_id in task_ids])
+        assignee_query = f"DELETE FROM assignees WHERE {where_clause}"
+        label_query = f"DELETE FROM task_labels WHERE {where_clause}"
+        database.delete(assignee_query, ())
+        database.delete(label_query, ())
 
     # Creating block and task delete queries
     block_query = "DELETE FROM blocks WHERE id = %s"
     task_query = "DELETE FROM tasks WHERE block_id = %s"
 
     # Deleting block and its related fields
-    database.delete(assignee_query, ())
-    database.delete(label_query, ())
     database.delete(block_query, values)
     database.delete(task_query, values)
